Assignment02/code/requirements.txt
cassandra-driver
pandas
kafka-python
watchdog
Flask
